AngularProject/src/main/webapp/app/entities/role-user-my-suffix/delete/role-user-my-suffix-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { RoleUserMySuffixService } from '../service/role-user-my-suffix.service';

import { RoleUserMySuffixDeleteDialogComponent } from './role-user-my-suffix-delete-dialog.component';

describe('RoleUserMySuffix Management Delete Component', () => {
  let comp: RoleUserMySuffixDeleteDialogComponent;
  let fixture: ComponentFixture<RoleUserMySuffixDeleteDialogComponent>;
  let service: RoleUserMySuffixService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RoleUserMySuffixDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(RoleUserMySuffixDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(RoleUserMySuffixDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(RoleUserMySuffixService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
